// src/app/auth/enums/roles.enum.ts
export enum RoleNames {
    ADMIN = 'administrador',
    ADMIN_CIARP = 'admin-ciarp',
    TEACHER = 'docente',
    CPD_SECRETARY = 'secretaria-cpd',
    CPD_PRESIDENT = 'presidente-cpd',
    CPD_MEMBER = 'integrante-cpd',
    CIARP_SECRETARY = 'secretaria-ciarp',
    CIARP_MEMBER = 'integrante-ciarp',
}